Ignore date/time values that cannot be parsed
CoreNLP has some very weird handling of times like "sunrise" or
"dawn" or "afternoon" that map to some tags.
While it would be nice to handle them (because "make me coffee
in the morning" is a reasonable thing to say to Sabrina), our
value representation is not rich enough to do that at the moment,
so just ignore them.