Cache lexicon lookups in memory
The parsing code will keep most or all derivations for a lexicon
lookup in memory, as part of the corresponding cell that invoked
ThingpediaLexiconFn. At that point, there is little point in
going streaming, so we should buffer the results of looking
up in the database, which immediately let's us cache the lookup
in memory.