Rework the way we typecheck channel <-> argument names
The previous way required getting all copies of arguments with
the same name, each for a different channel, and then filtering
out the copies from the wrong channel.
The problem comes when the same argument name is used in many
channels, as we get too many copies and we overflow the beam
size, meaning that we potentially miss some good argument names.
Furthermore, we can't use featurization to prefer some arguments
rather than others because they have the same canonical and type,
so they are virtually indistinguishable at the lexical level.

Instead, do the typechecking by retrieving the list of accepted
argument names and types for the channel when looking it up
in the lexicon, and matching against that in FilterInvalidInvocationFn.
This requires replacing NameValue with ChannelNameValue, a new
value type for trigger/action/query names.