APIServer: add support for serving multiple languages
Recognize a "locale" query parameter, and map it to a language
context, which is a set of query cache, parser and language model parameters.

For now the only recognized language is English, which is also
the fallback if the language is unrecognized, because the grammar,
lexicon and datasets don't support multiple instances with different
language tags.