Params: simplify locking
Right now locking is occasionally broken (for example getWeights()
returns the internal hash map outside the lock) and occasionally
too-fine grained (getWeight() continuously acquires and releases
the lock inside the hot FeatureVector.dotProduct() call)

Replace it with a read-write lock instead. The read lock is
acquired by the parser around the entire parse, while the write
lock is acquired by Params.update().
The rw lock is set to fair, which ensures eventually the write
lock is acquired, even with high read contention (which we expect).