Add more string features
Featurize POS tags for free-form strings. More often than not, free-form
strings are proper nouns (organizations, people, etc.). NER recognition is
spotty, but POS tagging consistenly recognizes them as NNP, which is useful.

We also add a feature about the string size, which makes sure we learn to
make the string as small as possible, to avoid running into a second string.