api: add an exact matching layer before the natural language analysis
If we know exactly the sentence we're asked to parse because we
have seen it in the database, we should skip the NL step and just
return the answer.
This allows user to train the NL with sentences outside the model,
for example "get tweets" to be "receive tweet on twitter", despite
never seeing the latter in the candidate list because it's not
produced by the lexicon.

The exact match layer is in-memory only. We don't expect we will
ever have so many utterances that this will become a problem.