Add a new way to do ThingTalk composition
Doing the parameter binding on the client side is too hacky because
too much info is lost at that point, and we don't have a full ML
infrastructure on the client. Instead, do it on the server by hacking
the grammar to generate action argument, trigger argument pairs
that are appended to the end of a rule.