ThingpediaLexicon: add feature aligning channel search keys to kinds
We want to recognize certain words as particularly indicative of
certain kinds, so that we don't go in the woods if the user uses
them.

(We can't use a single more generic "keyword == kind" feature because
in case of eg. "get xkcd", we have one derivation where we look up
"get" and one derivation where we look up "xkcd". The first derivation
does not have the feature, the second has it. They have the same value
so the feature is learnt as 0 weight.)