Simplify the process of adding values to actions/queries/triggers
Instead of doing it piecewise in three rules, have a semantic fn
that just generates all possible (name, operator) pairs that
bridge between an action and a value.

This greatly reduces the depth of actions with arguments (which
means we can cut on the floating parser depth), it reduces the
derivations that are featurized redundantly, and it cleans up
the code by removing ugly LambdaFormula manipulation.