JsonValueEvaluator: add a normalization step before comparing values
Jackson will deserialize JSON as a double or as an integer depending
on the .0 at the end of the number.
This is broken on so many levels (because doubles and integers
don't compare .equals()), but in order to correctly learn numbers
when the JSON is generated in JavaScript land (where doubles and integers
are one and the same) we need to do this.