Derivation: add nerUtterance to complement canonicalUtterance
NerUtterance is the canonical utterance with all instances of
NER values replaced with their tags.
These will be used for better lookups in the phrase table.

So far, ner utterances and canonical utterances are identical.
Semantic fns need to be updated to set meaningful values.

Example:

post on twitter with status "foo" --> post on twitter with status QUOTED_STRING
monitor if receive tweet with from @foo --> monitor if receive tweet with from USERNAME