Fix entity handling in overnight featurization not to depend on quoted string length
If we handle entities the old way we put a match for each word
that matches in the quoted string. With long quoted strings that's
a lot of matches, which means we prefer arguments that come later but
are longer and we lose by default.
This way, all quoted strings are worth the same.