Restructure starting scripts
Move all common options to a common configuration file, so that
we don't run the three modes with different options all the time.

There is now a single run.sh that takes a mode as the first
argument. Unfortunately, training is a bit special because it
runs with a temporary workdir, so it still has a wrapper script
that generates the configuration file.