Implement grammar constraints for decoding programs
Which avoids generating programs that do not parse.

We can still generate programs that don't typecheck, because weeding
those out is too hard. Also if the program hits the length limit
it can still fail to parse.

It's not the majesting improvement I was hoping for, but it's a
noticeable (10% when lucky) improvement still.