Rework how the output of decoding is computed
Instead of returning the sequence of logits and then doing the decode
again in numpy, leverage the fact that modern Tensorflow can return
the decoded tokens directly.

This will increase speed (as more computation can be done on the GPU)
and will also simplify adding grammar constraints back, as we only
need to care about constraining tensors.