wip: add beam search optimization-based aligner.
Not quite working yet, because it tries to use operations that
are not differentiable and Tensorflow (unsurprisingly) does not
like that.

We'll need to write a different loss, based fully on what the BSO
paper does. Which is so complicated that nobody on the Internet
is even trying to implement...