Add scripts to evaluate the encoder
Encoder evaluation looks at the confusion list generated by
run_test on the dev set and computes the distance to the average
encoder point for the gold and predicted program (taken as the
average over all training sentences for that program). It also
takes the form of 2-dimensional PCA in a plot, for visual.