Add the ability to break the connection from the decoder output to the input
Connecting the decoder output to the input of the next RNN cell
creates a language model, which is not what we want. Specifically,
we don't want to be biased towards certain combinations of inputs/outputs.

Also, cutting the connection decouples the size of output_embed_matrix
from the LSTM size, which is useful when the output_embed_matrix is used
as output layer.