ThingTalk: switch from DFA grammar to SLR grammar
ThingTalk is about to grow a lot of new operators, and the old
manually maintained DFA grammar is not going to cut it.
Instead, use the new SLR parser generator to generate a parser
for ThingTalk code. This will be used to produce a sequence of
shift-reduce actions that the NN will try to predict.