DotProductOutput: change the meaning
Doing a matmul against the embedding matrix and then a softmax loss
is not as meaningful as one might thing (actually, it's not meaningful
at all).

The matmul was meant as a proxy for cosine distance, but without
normalization that does not work. Instead, compute an L2 distance
between the RNN output and each entry in the embedding.
Do this both when computing the loss and when sampling.