ShiftReduceGrammar: use raw tokens for copying
We want to be able to copy words that are truly out-of-vocabulary,
ie they are not in input_words.txt and are vectorized as <unk>.
This is to reduce the vocabulary size that is exposed to the RNN,
and thus improve the quality of the model.

To do so, we must have vectorize_program and reconstruct_program
process the raw input sentences instead of the input vectors.