Beginning of the great change
The world is about to change.

This is the first commit to refactor and integrate t2t into
our codebase properly.

All our code is moved into a luinet/ python package (which by
the end of this series will be installable with distutils)

A command, luinet-datagen, is introduced, as a wrapper over
t2t-datagen with a few more options.
T2T problems now live in the luinet.tasks package.

This is enough to train a model for ThingTalk-with-quotes;
model changes will be needed for quote-free-ThingTalk.