Implement correct decoding of predicted programs
Using the standard t2t approach to decoding (an Encoder class
returned by Problem.feature_encoders()) does not work, because
at decoding time we need the input sentence to retrieve the
content of the span (the predicted program only includes
the indices of the span begin and end).

Instead, we run the decoding in the py_func that converts
the grammar actions to the linear predicted program (which we're running
anyway), conditioned on being in decode or eval mode.