Fix removing '*' from tokenized output
The latest version of almond-tokenizer adds * followed by a number
to all entity tokens (eg NUMBER_0*1 - the number indicates how
many original tokens are covered by the entity).
Make sure we remove that from the stuff we return to the client,
or the stuff we use to look up in the exact matcher, or we will
be very confused.