Fix reloading the problem in the same process
The server has the option to reload the model from disk, to support
continuous training. When this happens, it creates a new instance
of the problem and the model in a new Tensorflow Graph.

This is not compatible with messing with the t2t registry in
the problem, because the registry is a global resource.

Instead, skip the registry altogether and set the hparams to
a callback that will directly create the modality.