Add quote free support for the exact matcher
Now the exact matcher recognizes the part between quotes in the
program, finds the corresponding portion of the input sentence,
and turns it into a wildcard.

This is done on a word by word basis, so there is no ambiguity
and we don't need a complex parsing algorithm, which keeps the
exact matcher simple fast and reliable.

Fixes #10