ExactMatcher: allow ambiguous results
And return all of them

The purpose of this is to support multiple, unapproved devices
for the same services, with ambiguous sentences.

In that case, the client side will receive all results, and filter
out those that don't typecheck (as the client will only see one
unapproved devices + the approved ones).

In particular, this should help with the cat tutorial.