Add a script to read the metrics for the best model from evaluation logs
A script that walks the eval/*.tfevents files and prints the metrics
corresponding to the best model.
This script will be used primarily by the auto-trainer to get
a measure of model quality, but it is also useful to get the dev
statistics for a manually trained file, without parsing the text log files.