tasks/semantic_parsing: prune the dictionary by frequency
All words that appear less than a threshold number of times
in the dataset (50 by default) are given an unk word vector.

This helps when the model is deployed in production, as new
words would be true <unk>'s and the model would be confused
by having never seen them.